backend/requirements.txt
Flask>=2.3.3
Flask-CORS>=4.0.0
Pillow>=10.1.0
Werkzeug>=2.3.7
numpy>=1.24.0